Rename CalculateTaskGraphDescriptor and make its result an inner.
